+++ src/frontend/main.jsx
@@ -3,10 +3,10 @@ import {createRoot} from 'react-dom/client'
 import {BrowserRouter} from 'react-router-dom'
 import {Amplify} from 'aws-amplify'
 import {Authenticator, View} from '@aws-amplify/ui-react'
-import AppRoutes from "./js/routes";
+import AppRoutes from "./routes";
 
 
-const urls = ["http://localhost:5173"]
+const urls = ["http://localhost:4000"]
 const url = urls[0]
 
 console.log(`COGNITO_DOMAIN: ${process.env.COGNITO_DOMAIN}\nAWS_REGION: ${process.env.AWS_REGION}\nCOGNITO_USER_POOL_CLIENT_ID: ${process.env.COGNITO_USER_POOL_CLIENT_ID}\nCOGNITO_USER_POOL_ID: ${process.env.COGNITO_USER_POOL_ID}`)